Modular project structure.
