+++ package.json
@@ -27,8 +27,8 @@
   "files": [
     "lib"
   ],
-  "main": "lib/main/index.js",
-  "module": "lib/module/index.mjs",
+  "main": "lib",
+  "module": "lib/index.mjs",
   "engines": {
     "node": ">=7.6"
   },
@@ -51,9 +51,9 @@
     "prettier": "^1.9.2"
   },
   "scripts": {
-    "build": "npm run build:main && npm run build:module && prettier --write 'lib/**/*.{js,mjs}'",
-    "build:main": "babel src --out-dir lib/main",
-    "build:module": "MODULE=true babel src --out-dir lib/module --keep-file-extension",
+    "build": "npm run build:js && npm run build:mjs && prettier --write 'lib/**/*.{js,mjs}'",
+    "build:js": "babel src --out-dir lib",
+    "build:mjs": "MODULE=true babel src --out-dir lib --keep-file-extension",
     "lint": "eslint . --ext js,mjs",
     "fix": "npm run lint -- --fix && prettier --write '**/*.md'",
     "precommit": "lint-staged",
